+++ extras/tests/JsonDocument/ElementProxy.cpp
@@ -9,7 +9,7 @@ using namespace ARDUINOJSON_NAMESPACE;
 
 TEST_CASE("ElementProxy::add()") {
   DynamicJsonDocument doc(4096);
-  doc.addElement();
+  doc.add();
   ElementProxy<JsonDocument &> ep = doc[0];
 
   SECTION("add(int)") {
@@ -35,7 +35,7 @@ TEST_CASE("ElementProxy::add()") {
 
 TEST_CASE("ElementProxy::clear()") {
   DynamicJsonDocument doc(4096);
-  doc.addElement();
+  doc.add();
   ElementProxy<JsonDocument &> ep = doc[0];
 
   SECTION("size goes back to zero") {
@@ -95,7 +95,7 @@ TEST_CASE("ElementProxy::operator==()") {
 
 TEST_CASE("ElementProxy::remove()") {
   DynamicJsonDocument doc(4096);
-  doc.addElement();
+  doc.add();
   ElementProxy<JsonDocument &> ep = doc[0];
 
   SECTION("remove(int)") {
@@ -168,7 +168,7 @@ TEST_CASE("ElementProxy::set()") {
 
 TEST_CASE("ElementProxy::size()") {
   DynamicJsonDocument doc(4096);
-  doc.addElement();
+  doc.add();
   ElementProxy<JsonDocument &> ep = doc[0];
 
   SECTION("returns 0") {
@@ -190,7 +190,7 @@ TEST_CASE("ElementProxy::size()") {
 
 TEST_CASE("ElementProxy::memoryUsage()") {
   DynamicJsonDocument doc(4096);
-  doc.addElement();
+  doc.add();
   ElementProxy<JsonDocument &> ep = doc[0];
 
   SECTION("returns 0 for null") {