+++ extras/tests/JsonVariant/compare.cpp
@@ -10,7 +10,7 @@
 
 TEST_CASE("Compare JsonVariant with value") {
   StaticJsonDocument<256> doc;
-  JsonVariant a = doc.addElement();
+  JsonVariant a = doc.add();
 
   SECTION("null vs (char*)0") {
     char* b = 0;
@@ -38,8 +38,8 @@ TEST_CASE("Compare JsonVariant with value") {
 
 TEST_CASE("Compare JsonVariant with JsonVariant") {
   StaticJsonDocument<256> doc;
-  JsonVariant a = doc.addElement();
-  JsonVariant b = doc.addElement();
+  JsonVariant a = doc.add();
+  JsonVariant b = doc.add();
 
   SECTION("'abc' vs 'abc'") {
     a.set("abc");