+++ src/ArduinoJson/Array/ArrayImpl.hpp
@@ -11,12 +11,12 @@ namespace ARDUINOJSON_NAMESPACE {
 
 template <typename TArray>
 inline ArrayRef ArrayShortcuts<TArray>::createNestedArray() const {
-  return impl()->addElement().template to<ArrayRef>();
+  return impl()->add().template to<ArrayRef>();
 }
 
 template <typename TArray>
 inline ObjectRef ArrayShortcuts<TArray>::createNestedObject() const {
-  return impl()->addElement().template to<ObjectRef>();
+  return impl()->add().template to<ObjectRef>();
 }
 
 template <typename TArray>